Make it possible to refer to contentSize via WebViewProxy.
